Split building skia into building 3rd-party first then skia proper

+++ scripts/build_Linux.sh
@@ -4,6 +4,24 @@ export PATH=${PWD}/depot_tools:$PATH
 
 EXTRA_CFLAGS=""
 
+export CC=gcc
+export CXX=g++
+export AR=ar
+export CFLAGS="-Wno-deprecated-copy"
+export LDFLAGS="-lrt"
+
+# libicu.a is the largest 3rd-party; if it already exists, we simply run ninja
+# a 2nd time and exit.
+# Running ninja a 2nd-time is safe - it is no-ops if skia is already built too.
+# The 3rd-party below are built in size-order; we built libicu last to signal
+# having built most of them.
+if [[ -f "skia/out/Release/libicu.a" ]] ; then
+    cd skia && \
+        ninja -C out/Release && \
+        cd ..
+    exit 0
+if
+
 if [[ $(uname -m) == "aarch64" ]]; then
     # Install ninja for aarch64
     yum -y install epel-release && \
@@ -25,17 +43,7 @@ if [[ $EUID -eq 0 ]]; then
         rm -rf /var/cache/yum
 fi
 
-if [[ $(uname -m) == "aarch64" ]] && [[ $CI_SKIP_BUILD == "true" ]]; then
-    # gn and skia already built in a previous job
-    exit 0
-fi
-
 # Build gn
-export CC=gcc
-export CXX=g++
-export AR=ar
-export CFLAGS="-Wno-deprecated-copy"
-export LDFLAGS="-lrt"
 git clone https://gn.googlesource.com/gn && \
     cd gn && \
     git checkout fe330c0ae1ec29db30b6f830e50771a335e071fb && \
@@ -61,5 +69,12 @@ skia_use_system_freetype2=false
 extra_cflags_cc=[\"-frtti\"]
 extra_ldflags=[\"-lrt\"]
 " && \
-    ninja -C out/Release && \
+    ninja -C out/Release \
+          third_party/freetype2 \
+          third_party/libwebp \
+          third_party/dng_sdk \
+          third_party/harfbuzz \
+          third_party/icu && \
+    ( ( [[ $(uname -m) == "aarch64" ]] && echo "On aarch64 - Please run me again!" ) || \
+          ninja -C out/Release ) && \
     cd ..
